requirements.txt
beautifulsoup4==4.11.1
langchain==0.3.14
langchain_astradb==0.5.2
langchain_community==0.3.14
langchain_core==0.3.29
langchain_huggingface==0.1.2
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.41.1
torch==2.5.1
transformers==4.47.1